Dump performance metrics daily
