+++ lobby/SQLiteConnection.cpp
@@ -197,3 +197,15 @@ void SQLiteInstance::setBusyTimeout(int milliseconds)
 {
 	sqlite3_busy_timeout(m_connection, milliseconds);
 }
+
+void SQLiteInstance::printMemoryStats() const
+{
+	sqlite3_int64 currentUsage = sqlite3_memory_used();
+	sqlite3_int64 peakSinceLastCall = sqlite3_memory_highwater(1);
+
+	logGlobal->info("SQLite instance memory usage: %d B (%d MB). Peak was %d B (%d MB)",
+					currentUsage, currentUsage / 1024 / 1924,
+					peakSinceLastCall, peakSinceLastCall / 1024 / 1924);
+
+	// more detailed stats can be acquired using sqlite3_status
+}
