+++ lobby/SQLiteConnection.h
@@ -118,6 +118,7 @@ public:
 
 	SQLiteStatementPtr prepare(const std::string & statement);
 	void setBusyTimeout(int milliseconds);
+	void printMemoryStats() const;
 
 private:
 	explicit SQLiteInstance(sqlite3 * connection);
